refactor: move card service to parent component
